requirements.txt
attrs==21.4.0
blinker==1.4
click==8.1.3
Flask==2.1.2
Flask-Mail==0.9.1
greenlet==1.1.2
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pdfkit==1.0.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
sha256==0.3
SQLAlchemy==1.4.36
tomli==2.0.1
Werkzeug==2.1.2
zipp==3.8.0
